requirements.txt
asgiref==3.8.1
beautifulsoup4==4.12.3
Brotli==1.1.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Django==5.1
djangorestframework==3.15.2
gunicorn==23.0.0
h11==0.14.0
idna==3.7
packaging==24.1
requests==2.32.3
soupsieve==2.6
sqlparse==0.5.1
urllib3==2.2.2
uvicorn==0.30.6
whitenoise==6.7.0
